@@ -11,24 +11,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.2",
+    "@angular/animations": "~10.1.3",
     "@angular/cdk": "^10.2.2",
-    "@angular/common": "~10.0.2",
-    "@angular/compiler": "~10.0.2",
-    "@angular/core": "~10.0.2",
-    "@angular/forms": "~10.0.2",
+    "@angular/common": "~10.1.3",
+    "@angular/compiler": "~10.1.3",
+    "@angular/core": "~10.1.3",
+    "@angular/forms": "~10.1.3",
     "@angular/material": "^10.2.2",
-    "@angular/platform-browser": "~10.0.2",
-    "@angular/platform-browser-dynamic": "~10.0.2",
-    "@angular/router": "~10.0.2",
+    "@angular/platform-browser": "~10.1.3",
+    "@angular/platform-browser-dynamic": "~10.1.3",
+    "@angular/router": "~10.1.3",
+    "angular-material": "^1.2.1",
     "rxjs": "~6.6.3",
+    "rxjs-compat": "^6.6.3",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.1001.3",
     "@angular/cli": "~10.1.3",
-    "@angular/compiler-cli": "~10.0.2",
+    "@angular/compiler-cli": "~10.1.3",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
