@@ -15,21 +15,21 @@ import { NavigationbarComponent } from './navigationbar/navigationbar.component'
 import { AddtourComponent } from './addtour/addtour.component';
 import { PersonsComponent } from './persons/persons.component';
 import { AddtnavbarComponent } from './addtour/addtnavbar/addtnavbar.component';
-import {MatFormFieldModule} from "@angular/material/form-field";
-import {MatDatepickerModule} from "@angular/material/datepicker";
+import {MatFormFieldModule} from '@angular/material/form-field';
+import {MatDatepickerModule} from '@angular/material/datepicker';
 import {MatNativeDateModule} from '@angular/material/core';
-import {MatInputModule} from "@angular/material/input";
-import {MatSelectModule} from "@angular/material/select";
-import {FormsModule, ReactiveFormsModule} from "@angular/forms";
+import {MatInputModule} from '@angular/material/input';
+import {MatSelectModule} from '@angular/material/select';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { DayComponent } from './addtour/day/day.component';
-import {MatCardModule} from "@angular/material/card";
-import {MatSnackBarModule} from "@angular/material/snack-bar";
-import {MatAutocompleteModule} from "@angular/material/autocomplete";
+import {MatCardModule} from '@angular/material/card';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import { TourComponent } from './addtour/day/tour/tour.component';
-import {HttpClient, HttpClientModule} from "@angular/common/http";
+import {HttpClient, HttpClientModule} from '@angular/common/http';
 import { FormComponent } from './addtour/day/tour/form/form.component';
-import {MatGridListModule} from "@angular/material/grid-list";
-import {MatTableModule} from "@angular/material/table";
+import {MatGridListModule} from '@angular/material/grid-list';
+import {MatTableModule} from '@angular/material/table';
 import { PersonstableComponent } from './persons/personstable/personstable.component';
 import { MatPaginatorModule } from '@angular/material/paginator';
 import { MatSortModule } from '@angular/material/sort';
